Add support for enumerating available contexts using a new API
This allows applications to discover the available contexts and present a
list to users where they can choose which context they want to use.

The previous attempt at implementing context discovery resulted in an
API that did not allow us to support dynamic connection and
disconnection of IIO contexts.

This new API is now composed of only three functions:
- iio_create_scan_context(), that takes a function callback as argument,
  which will be called everytime a new IIO context connects or
  disconnects. It returns a pointer to an opaque 'iio_scan_context'
  structure.
- iio_scan_context_destroy(), to free all the resources allocated by the
  previous function;
- iio_scan_context_poll(), which will process connection/disconnection
  events of the compatible backends, and call the provided callback
  function when it applies.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>