Revert "IIOD: Refill the buffer when no clients, if we had clients just before"
This optimization was useful when it was introduced, in May 2014.
It was based on the assumption that if the RW thread doesn't have
readers, but had readers in the previous iteration, then it's highly
likely that some clients will ask for data in this iteration as well,
while we are refilling the buffer.

However, this introduced a corner case that caused the IIOD to drop one
buffer of samples, if no clients asked for data during the refill.
This typically happens when using the high-speed buffer interface in the
local backend, and with a sample rate higher than what the network can
deliver, since then refilling the buffer hardly takes any time.
This went unnoticed, mostly because in that typical use case, a lot of
samples are dropped in the kernel anyway, as the DMA copies faster than
we can refill.

This revert will have implications on the performance. The network
communication will suffer more from the latency, as we don't refill and
handle clients in parallel anymore. This is an issue that we will have
to handle differently in the future.

This reverts commit 846898575b12570560680815f8538984a0265132.