local: device_check_ready(): Handle EINTR during poll()
When waiting for the device to become ready we need to retry the poll
operation if it is interrupted by a signal, otherwise a signal will abort
blocking IO operations, which is not the expected behaviour with libiio.

The tricky part is that for correct behavior the timeout needs to be
adjusted. Otherwise signals could prevent the operation from ever timing
out. Use the monotonic clock to get the timestamp at the start of the
operation and calculate the timeout that is passed to poll() based on the
amount of time that has already elapsed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>